requirements.txt
Django
gunicorn
django-heroku
django-bootstrap-v5
django-allauth
requests
apscheduler
icalendar
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
whitenoise